# Conflicts:
#	docs/LICENSE_OF_DEPENDENCIES.md
#	go.mod

+++ plugins/inputs/kernel_vmstat/kernel_vmstat.go
@@ -1,3 +1,4 @@
+//go:build linux
 // +build linux
 
 package kernel_vmstat
@@ -45,7 +46,7 @@ func (k *KernelVmstat) Gather(acc telegraf.Accumulator) error {
 				return err
 			}
 
-			fields[string(field)] = int64(m)
+			fields[string(field)] = m
 		}
 	}
